requirements.txt
cryptography
azure-functions
azure-storage-blob
azure-identity
requests
pandas
numpy
PyPDF2